+++ src/vw/Image/ImageResourceOpenCV.cc
@@ -29,7 +29,7 @@ bool ImageResourceOpenCV::contiguous_roi(const BBox2i& bbox) const {
 
   return m_matrix->isContinuous()
          && bbox.min().x() == 0
-         && bbox.width() == m_format.cols;
+         && uint32(bbox.width()) == m_format.cols;
 }
 
 ImageFormat ImageResourceOpenCV::identify() const {
@@ -73,9 +73,9 @@ ImageResourceOpenCV::ImageResourceOpenCV(boost::shared_ptr<cv::Mat> matrix)
 }
 
 void ImageResourceOpenCV::read( ImageBuffer const& dst_buf, BBox2i const& bbox ) const {
-  VW_ASSERT(dst_buf.format.cols == bbox.width() && dst_buf.format.rows == bbox.height(),
+  VW_ASSERT(dst_buf.format.cols == uint32(bbox.width()) && dst_buf.format.rows == uint32(bbox.height()),
       LogicErr()    << VW_CURRENT_FUNCTION << ": Destination buffer has wrong dimensions!" );
-  VW_ASSERT(bbox.min().x() >= 0 && bbox.min().y() >= 0 && bbox.max().x() <= m_format.cols && bbox.max().y() <= m_format.rows,
+  VW_ASSERT(bbox.min().x() >= 0 && bbox.min().y() >= 0 && uint32(bbox.max().x()) <= m_format.cols && uint32(bbox.max().y()) <= m_format.rows,
       ArgumentErr() << VW_CURRENT_FUNCTION << ": Bounding box must be inside matrix.");
 
   ImageFormat src_format = m_format;
@@ -99,9 +99,9 @@ void ImageResourceOpenCV::read( ImageBuffer const& dst_buf, BBox2i const& bbox )
 }
 
 void ImageResourceOpenCV::write( ImageBuffer const& src_buf, BBox2i const& bbox ) {
-  VW_ASSERT(src_buf.format.cols == bbox.width() && src_buf.format.rows == bbox.height(),
+  VW_ASSERT(src_buf.format.cols == uint32(bbox.width()) && src_buf.format.rows == uint32(bbox.height()),
       LogicErr()    << VW_CURRENT_FUNCTION << ": Source buffer has wrong dimensions!" );
-  VW_ASSERT(bbox.min().x() >= 0 && bbox.min().y() >= 0 && bbox.max().x() <= m_format.cols && bbox.max().y() <= m_format.rows,
+  VW_ASSERT(bbox.min().x() >= 0 && bbox.min().y() >= 0 && uint32(bbox.max().x()) <= m_format.cols && uint32(bbox.max().y()) <= m_format.rows,
       ArgumentErr() << VW_CURRENT_FUNCTION << ": Bounding box must be inside matrix.");
 
   ImageFormat dst_format = m_format;
